refactor: move PlanSession and PlanExecutor to their own module
run_engine.py was 3,484 lines and held three classes with quite
different jobs. The session and the executor, and the pieces only they
use, move to plan_executor.py. run_engine.py keeps the RunEngine and
the forwarding shim, and is now 1,222 lines.

The Dispatcher and the bluesky event loop helpers had to come too. The
session owns a Dispatcher, and the executor asks the loop helpers which
thread it is on, so leaving them behind would have made the two modules
import each other. Moving the event loop helpers also keeps there being
exactly one _bluesky_event_loop global rather than one per module.

run_engine.py re-exports everything it used to define, with an explicit
__all__ so the re-exports cannot be tidied away as unused. That covers
the names other modules take from it, including DocumentNames, which
bluesky.callbacks.zmq imports from run_engine rather than from
event_model.

_state_locked is deleted. Its only caller was _clear_call_cache, which
became a call to _new_executor two commits ago.
