+++ src/bluesky/run_engine.py
@@ -1,54 +1,43 @@
 import asyncio
 import concurrent
-import copy
 import functools
 import inspect
-import json
-import sys
 import threading
 import typing
-import weakref
-from collections import ChainMap, defaultdict, deque
-from collections.abc import Callable, MutableMapping
 from contextlib import ExitStack
-from dataclasses import dataclass
-from datetime import datetime
-from enum import Enum
-from inspect import iscoroutine
-from itertools import count
-from logging import LoggerAdapter
 from warnings import warn
 
-import event_model
-from event_model import DocumentNames
-from opentelemetry import trace
-from opentelemetry.trace import Span
-
 from bluesky._vendor.super_state_machine.errors import TransitionError
-from bluesky._vendor.super_state_machine.extras import PropertyMachine
-from bluesky._vendor.super_state_machine.machines import StateMachine
 
-from .bundlers import RunBundler, maybe_await
-from .log import ComposableLogAdapter, logger, msg_logger, state_logger
-from .protocols import (
-    Flyable,
-    Locatable,
-    Movable,
-    Pausable,
-    Preparable,
-    Readable,
-    Stageable,
-    Status,
-    Stoppable,
-    SyncOrAsync,
-    T,
-    Triggerable,
-    check_supports,
+from .bundlers import RunBundler
+from .log import ComposableLogAdapter, logger
+
+# Everything this module used to define now lives in plan_executor, and is
+# exported from there. These are imported, and deliberately left out of the
+# __all__ below, so that `from bluesky.run_engine import Dispatcher` and its
+# like keep working for code written before the split.
+from .plan_executor import (  # noqa: F401
+    NO_PLAN_RETURN,
+    UNCACHEABLE_COMMANDS,
+    Dispatcher,
+    DocumentNames,
+    LoggingPropertyMachine,
+    PlanExecutor,
+    PlanSession,
+    RunEngineMetadata,
+    RunEngineResult,
+    RunEngineStateMachine,
+    WaitForTimeoutError,
+    _ensure_event_loop_running,
+    autoawait_in_bluesky_event_loop,
+    call_in_bluesky_event_loop,
+    default_scan_id_source,
+    get_bluesky_event_loop,
+    in_bluesky_event_loop,
+    set_bluesky_event_loop,
 )
-from .tracing import tracer
+from .protocols import SyncOrAsync
 from .utils import (
-    AsyncInput,
-    CallbackRegistry,
     DefaultDuringTask,
     DuringTask,
     FailedPause,
@@ -63,2101 +52,55 @@ from .utils import (
     RunEngineInterrupted,
     SigintHandler,
     Subscribers,
-    ensure_generator,
-    normalize_subs_input,
-    sanitize_np,
     single_gen,
-    warn_if_msg_args_or_kwargs,
 )
 
-_SPAN_NAME_PREFIX = "Bluesky RunEngine"
-
-
-class _NoPlanReturn:
-    """The type of :data:`NO_PLAN_RETURN`."""
-
-    def __repr__(self) -> str:
-        return "NO_PLAN_RETURN"
-
-
-#: Returned in place of a plan's return value when the plan did not run to
-#: completion, and so never returned one. Distinguishable from a plan that
-#: completed and returned ``None``.
-NO_PLAN_RETURN = _NoPlanReturn()
-
-#: Commands that must not be replayed when rewinding to a checkpoint, either
-#: because they act on the RunEngine itself or because they are not idempotent.
-UNCACHEABLE_COMMANDS = frozenset(
-    {
-        "pause",
-        "subscribe",
-        "unsubscribe",
-        "stage",
-        "unstage",
-        "monitor",
-        "unmonitor",
-        "open_run",
-        "close_run",
-        "install_suspender",
-        "remove_suspender",
-        "_start_suspender",
-    }
-)
+# What this module defines, plus the names it has always passed through from
+# bluesky.utils and event_model. The names that moved to plan_executor are
+# exported from there, not here.
+__all__ = [
+    "MAX_DEPTH_EXCEEDED_ERR_MSG",
+    "PAUSE_MSG",
+    "DocumentNames",
+    "FailedPause",
+    "FailedStatus",
+    "IllegalMessageSequence",
+    "InvalidCommand",
+    "Msg",
+    "NoReplayAllowed",
+    "PlanHalt",
+    "RequestAbort",
+    "RequestStop",
+    "RunEngine",
+    "RunEngineInterrupted",
+    "TransitionError",
+]
 
 
 class _RunEnginePanic(Exception): ...
 
 
-class WaitForTimeoutError(TimeoutError): ...
-
-
-@dataclass
-class RunEngineResult:
-    """
-    Information about the plan that was run
-
-    Attributes
-    ----------
-    run_start_uids : list
-        A list of the UIDs generated during the plan (if any)
-    plan_result :
-        The return value of the top-level plan that was run
-    exit_status : str
-    interrupted : bool
-        True if the plan was halted, stopped or aborted
-    reason : str
-        A text description of the reason why the plan was aborted (if aborted)
-    exception :
-        The exception generated by the plan, if any
-    """
-
-    run_start_uids: tuple[str, ...]
-    plan_result: typing.Any
-    exit_status: str
-    interrupted: bool
-    reason: str
-    exception: Exception | None
-
-
-class RunEngineStateMachine(StateMachine):
-    """
-
-    Attributes
-    ----------
-    is_idle
-        State machine is in its idle state
-    is_running
-        State machine is in its running state
-    is_paused
-        State machine is paused.
-    """
-
-    class States(Enum):
-        """state.name = state.value"""
-
-        IDLE = "idle"
-
-        RUNNING = "running"
-
-        PAUSING = "pausing"
-        PAUSED = "paused"
-
-        HALTING = "halting"
-        STOPPING = "stopping"
-        ABORTING = "aborting"
-
-        SUSPENDING = "suspending"
-
-        PANICKED = "panicked"
-
-        @classmethod
-        def states(cls):
-            return [state.value for state in cls]
-
-    class Meta:
-        allow_empty = False
-        initial_state = "idle"
-        transitions = {
-            # Notice that 'transitions' and 'named_transitions' have
-            # opposite to <--> from structure.
-            # from_state : [valid_to_states]
-            "idle": ["running", "panicked"],
-            "running": ["idle", "pausing", "halting", "stopping", "aborting", "suspending", "panicked"],
-            "pausing": ["paused", "idle", "halting", "aborting", "panicked"],
-            "suspending": ["running", "halting", "aborting", "panicked"],
-            "paused": ["idle", "running", "halting", "stopping", "aborting", "panicked"],
-            "halting": ["idle", "panicked"],
-            "stopping": ["idle", "panicked"],
-            "aborting": ["idle", "panicked"],
-            "panicked": [],
-        }
-        named_checkers = [
-            ("can_pause", "pausing"),
-        ]
-
-
-class LoggingPropertyMachine(PropertyMachine):
-    """expects object to have a `log` attribute
-    and a `state_hook` attribute that is ``None`` or a callable with signature
-    ``f(value, old_value)``"""
-
-    def __init__(self, machine_type):
-        super().__init__(machine_type)
-
-    def __set__(self, obj, value):
-        own = type(obj)
-        old_value = self.__get__(obj, own)
-        with obj._state_lock:
-            super().__set__(obj, value)
-        value = self.__get__(obj, own)
-        tags = {"old_state": old_value, "new_state": value, "RE": self}
-
-        state_logger.info("Change state on %r from %r -> %r", obj, old_value, value, extra=tags)
-        if obj.state_hook is not None:
-            obj.state_hook(value, old_value)
-
-    def __get__(self, instance, owner):
-        if instance is None:
-            return super().__get__(instance, owner)
-        with instance._state_lock:
-            return super().__get__(instance, owner)
-
-
-RunEngineMetadata = MutableMapping[str, typing.Any]
-
-
-def default_scan_id_source(md: RunEngineMetadata) -> SyncOrAsync[int]:
-    return md.get("scan_id", 0) + 1
-
-
-def _state_locked(func):
-    @functools.wraps(func)
-    def inner(self, *args, **kwargs):
-        with self._state_lock:
-            return func(self, *args, **kwargs)
-
-    return inner
-
-
-def _default_event_loop() -> asyncio.AbstractEventLoop:
-    """The loop to use when none was given: this one, or the RunEngine's."""
-    try:
-        return asyncio.get_running_loop()
-    except RuntimeError:
-        pass
-    loop = get_bluesky_event_loop()
-    if loop is None:
-        raise RuntimeError(
-            "No event loop to run plans on. Either construct this from a "
-            "coroutine, so that there is a running loop to adopt, or pass "
-            "one in as loop=."
-        )
-    return loop
-
-
-class PlanSession:
-    """The environment that plans are executed in.
-
-    A session holds everything that outlives any single plan: the persistent
-    metadata, the document routing, the suspenders, the hooks, and the state
-    machine. It does not execute anything itself; a
-    :class:`PlanExecutor` does that, and one session may be used by many
-    executors in turn.
-
-    A `RunEngine` composes a session with the machinery needed to drive it
-    from a terminal on the main thread. A session needs no `RunEngine`, so it
-    can also be used directly from code that is already running in an asyncio
-    event loop, such as a headless data acquisition service.
-
-    Parameters
-    ----------
-    md : MutableMapping[str, Any], optional
-        The default is a standard Python dictionary, but fancier objects can
-        be used to store long-term history and persist it between sessions.
-        Any object adhering to the MutableMapping Protocol will work.
-
-    loop : asyncio event loop, optional
-        The loop plans will be executed on. Defaults to the running loop, or
-        to the loop a `RunEngine` has already established.
-
-    preprocessors : list, optional
-        Generator functions that take in a plan (generator instance) and
-        modify its messages on the way out. Functions are composed in order:
-        the preprocessors ``[f, g]`` are applied like ``f(g(plan))``.
-
-    md_validator : callable, optional
-        A function that raises and prevents starting a run if it deems the
-        metadata to be invalid or incomplete.
-
-    md_normalizer : callable, optional
-        A function that, like ``md_validator``, raises to prevent starting a
-        run, but which returns the normalized metadata if it succeeds.
-
-    scan_id_source : callable, optional
-        A (possibly async) function used to calculate ``scan_id``.
-
-    log : logging.LoggerAdapter, optional
-        Where this session and its executors log to.
-
-    on_pause : callable, optional
-        Called with no arguments when an executor reaches a paused resting
-        state. A `RunEngine` uses this to release the main thread; a headless
-        caller has no thread to release and can leave it unset.
-
-    Attributes
-    ----------
-    state
-        {'idle', 'running', 'paused'}
-
-    dispatcher
-        The `Dispatcher` documents are emitted through.
-
-    suspenders
-        Read-only collection of `bluesky.suspenders.SuspenderBase` objects.
-    """
-
-    _state = LoggingPropertyMachine(RunEngineStateMachine)
-
-    def __init__(
-        self,
-        md: RunEngineMetadata | None = None,
-        *,
-        loop: asyncio.AbstractEventLoop | None = None,
-        preprocessors: list | None = None,
-        md_validator: typing.Callable | None = None,
-        md_normalizer: typing.Callable | None = None,
-        scan_id_source: typing.Callable[[RunEngineMetadata], SyncOrAsync[int]] = default_scan_id_source,
-        log: LoggerAdapter | None = None,
-        on_pause: typing.Callable[[], None] | None = None,
-        run_bundler_cls: type[RunBundler] = RunBundler,
-        run_engine_cls: type | None = None,
-    ):
-        if loop is None:
-            loop = _default_event_loop()
-        self._loop = loop
-
-        # Guards transitions of the state machine, which a RunEngine writes to
-        # from the main thread as well as from the loop. Nothing else in a
-        # session or an executor takes a lock.
-        self._state_lock = threading.RLock()
-
-        self.log = log if log is not None else ComposableLogAdapter(logger, {"RE": self})
-
-        if md is None:
-            md = {}
-        self.md = md
-        self.md.setdefault("versions", {})
-
-        try:
-            import ophyd
-
-            self.md["versions"]["ophyd"] = ophyd.__version__
-        except ImportError:
-            self.log.debug("Failed to import ophyd.")
-
-        try:
-            import ophyd_async
-
-            self.md["versions"]["ophyd_async"] = ophyd_async.__version__
-        except ImportError:
-            self.log.debug("Failed to import ophyd_async.")
-
-        from ._version import __version__
-
-        self.md["versions"]["bluesky"] = __version__
-        self.md["versions"]["event_model"] = event_model.__version__
-
-        self.run_bundler_cls = run_bundler_cls
-        self.run_engine_cls = run_engine_cls
-        self.preprocessors = preprocessors if preprocessors is not None else []
-        self.md_validator = md_validator if md_validator is not None else _default_md_validator
-        self.md_normalizer = md_normalizer if md_normalizer is not None else _default_md_normalizer
-        self.scan_id_source = scan_id_source
-
-        self.msg_hook: typing.Callable | None = None
-        self.state_hook: typing.Callable | None = None
-        self.waiting_hook: typing.Callable | None = None
-        self.record_interruptions = False
-        self.on_pause = on_pause
-        self._require_stream_declaration = False
-
-        # Whether messages may be replayed on a rewind. A user setting that
-        # outlives any one plan, although a plan may change it for its own
-        # duration with Msg('rewindable', ...).
-        self.rewindable_flag: bool = True
-
-        self._suspenders: set[typing.Any] = set()
-
-        # Commands the user has added or removed. An executor composes these
-        # over its own built-ins, so that registrations survive the executor
-        # that was current when they were made.
-        self._registered_commands: dict[str, typing.Callable] = {}
-        self._unregistered_commands: set[str] = set()
-
-        # The executor currently running, or last to run, a plan. Suspenders
-        # and other session-level callers reach the live plan through this.
-        self.executor: PlanExecutor | None = None
-
-        # public dispatcher for callbacks
-        self.dispatcher = Dispatcher()
-        self.ignore_exceptions = False
-
-    @property
-    def loop(self) -> asyncio.AbstractEventLoop:
-        """The event loop plans are executed on."""
-        return self._loop
-
-    @property
-    def state(self):
-        """One of {'idle', 'running', 'paused'}."""
-        return self._state
-
-    @property
-    def ignore_exceptions(self):
-        return self.dispatcher.ignore_exceptions
-
-    @ignore_exceptions.setter
-    def ignore_exceptions(self, val):
-        self.dispatcher.ignore_exceptions = val
-
-    @property
-    def suspenders(self):
-        """Read-only collection of installed suspenders."""
-        return tuple(self._suspenders)
-
-    def subscribe(self, func, name="all"):
-        """Register a callback function to consume documents.
-
-        See :meth:`RunEngine.subscribe`.
-        """
-        return self.dispatcher.subscribe(func, name)
-
-    def unsubscribe(self, token):
-        """Unregister a callback function by its integer ID."""
-        return self.dispatcher.unsubscribe(token)
-
-    def new_executor(self, subs=None):
-        """Build the executor for the next plan and adopt it as the current one.
-
-        Replacing the executor is how the state of the previous plan is
-        cleared, so this is also where that plan's temporary subscriptions are
-        torn down: they last until the next plan starts, not until their own
-        finishes.
-
-        Parameters
-        ----------
-        subs : callable, list, or dict, optional
-            Subscriptions that last only as long as the plan the new executor
-            is about to run. Same forms as :meth:`RunEngine.__call__` accepts.
-        """
-        if self.executor is not None:
-            self.executor.clear_temp_subscriptions()
-        self.executor = PlanExecutor(self)
-        for name, funcs in normalize_subs_input(subs).items():
-            for func in funcs:
-                self.executor.add_temp_subscription(self.subscribe(func, name))
-        return self.executor
-
-    def emit_sync(self, name, doc):
-        "Process blocking callbacks and schedule non-blocking callbacks."
-        self.dispatcher.process(name, doc)
-
-    async def emit(self, name, doc):
-        self.emit_sync(name, doc)
-
-    def _call_waiting_hook(self, *args, **kwargs):
-        """Notify the waiting hook, if one is set."""
-        if self.waiting_hook is not None:
-            self.waiting_hook(*args, **kwargs)
-
-    def _notify_paused(self):
-        """Announce that an executor has reached a paused resting state."""
-        if self.on_pause is not None:
-            self.on_pause()
-
-    def install_suspender(self, suspender):
-        """Install a suspender, which can suspend and resume execution."""
-        self._suspenders.add(suspender)
-        suspender.install(self)
-
-    def remove_suspender(self, suspender):
-        """Uninstall a suspender."""
-        if suspender in self._suspenders:
-            suspender.remove()
-        self._suspenders.discard(suspender)
-
-    def clear_suspenders(self):
-        """Uninstall all suspenders."""
-        for suspender in self.suspenders:
-            self.remove_suspender(suspender)
-
-    def _announce_suspend(self):
-        """Tell the user a suspension is starting."""
-        print("Suspending....To get prompt hit Ctrl-C twice to pause.")
-        ts = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        print(f"Suspension occurred at {ts}.")
-
-    def request_suspend(self, fut, *, pre_plan=None, post_plan=None, justification=None, announce=True):
-        """Ask the plan in progress to suspend until ``fut`` is finished.
-
-        Must be called on the event loop; a `RunEngine` offers a thread-safe
-        version of this for callers that are not on it.
-
-        ``announce`` is for callers that have already told the user, so that
-        the message reaches them when they asked rather than whenever the
-        loop gets to it.
-        """
-        if announce:
-            self._announce_suspend()
-        self.loop.create_task(  # noqa: RUF006
-            self.executor.request_suspend(fut, pre_plan=pre_plan, post_plan=post_plan, justification=justification)
-        )
-
-
-class PlanExecutor:
-    """Executes one plan, which may contain any number of runs.
-
-    An executor owns everything that belongs to the execution of a single
-    plan: the stack of generators being worked off, the messages cached for
-    rewinding, the devices seen and staged, and the status objects being
-    waited on. It is built for a plan and discarded after it, so clearing
-    those caches is a matter of building a new one.
-
-    It lives entirely on its session's event loop. It holds no locks and no
-    threading primitives, and none of its methods may be called from another
-    thread; a `RunEngine` marshals everything the main thread asks for onto
-    the loop before it arrives here.
-
-    Parameters
-    ----------
-    session : PlanSession
-        The environment to execute in: metadata, document routing, hooks and
-        the state machine. Many executors may share one session, in turn.
-    """
-
-    def __init__(self, session: PlanSession):
-        self._session = session
-
-        # When cleared, run() will pause until it is set again.
-        self._run_permit = asyncio.Event()
-        self._run_permit.set()
-        # When set, done callbacks from status objects belonging to this plan
-        # stop reporting failures: the plan they belong to is over.
-        self._pardon_failures = asyncio.Event()
-
-        self._task: asyncio.Task | None = None  # the task running this plan
-        self._plan: typing.Any = None  # the plan, kept for metadata introspection
-        self._plan_stack: deque[typing.Any] = deque()  # generators to work off of
-        self._response_stack: deque[typing.Any] = deque()  # responses to send into them
-        # Processed msgs, for rewinding. None once a 'clear_checkpoint' has
-        # made this plan unrewindable, which is what `resumable` tests for.
-        self._msg_cache: deque[typing.Any] | None = deque()
-
-        self._run_bundlers: dict[typing.Any, RunBundler] = {}  # open run -> bundler
-        self._metadata_per_call: dict[typing.Any, typing.Any] = {}  # md for every run
-        self.run_start_uids: list[typing.Any] = []  # RunStart uids generated
-        self._run_tracing_spans: list[Span] = []  # open tracing spans
-
-        self._staged: set[typing.Any] = set()  # staged, not yet unstaged
-        self._objs_seen: set[typing.Any] = set()  # every object seen in a Msg
-        self._movable_objs_touched: set[typing.Any] = set()  # everything we 'set'
-        self._groups: defaultdict[str, set[Callable[[], asyncio.Future]]] = defaultdict(set)
-        self._status_objs: defaultdict[typing.Any, set[typing.Any]] = defaultdict(set)
-        self._seen_wait_and_move_on_keys: set[typing.Any] = set()
-
-        # An exception instance or class, to be raised into the plan.
-        self.exception: typing.Any = None
-        self.interrupted: bool = False  # paused, aborted or failed
-        self.exit_status: str = "success"  # optimistic default
-        self.reason: str = ""  # reason for an abort
-        self._deferred_pause_requested: bool = False  # pause at next 'checkpoint'
-
-        # Tokens for subscriptions that last only as long as this plan. They
-        # are tokens into the session's dispatcher, so that permanent and
-        # temporary callbacks stay in one registry, firing in one pass in
-        # subscription order; only the tokens belong to the plan.
-        self._temp_callback_ids: set[typing.Any] = set()
-
-        self.command_registry: dict[str, typing.Callable] = {}
-        self.rebuild_command_registry()
-
-    def add_temp_subscription(self, token):
-        """Record a token as belonging to this plan."""
-        self._temp_callback_ids.add(token)
-
-    def drop_temp_subscription(self, token):
-        """Unsubscribe a temporary subscription and forget its token."""
-        self._session.unsubscribe(token)
-        self._temp_callback_ids.remove(token)
-
-    def clear_temp_subscriptions(self):
-        """Unsubscribe every temporary subscription this plan made."""
-        for token in self._temp_callback_ids:
-            self._session.unsubscribe(token)
-        self._temp_callback_ids.clear()
-
-    @property
-    def session(self) -> PlanSession:
-        """The environment this plan is being executed in."""
-        return self._session
-
-    @property
-    def resumable(self) -> bool:
-        "i.e., can the plan in progress be rewound"
-        return self._msg_cache is not None
-
-    @property
-    def rewindable_flag(self) -> bool:
-        """Whether messages may be replayed. Held by the session, which
-        outlives this plan, since it is a user setting rather than plan
-        state."""
-        return self._session.rewindable_flag
-
-    @rewindable_flag.setter
-    def rewindable_flag(self, value: bool) -> None:
-        self._session.rewindable_flag = value
-
-    def permit_run(self) -> None:
-        """Allow run() to proceed. Must be called on the event loop."""
-        self._run_permit.set()
-
-    def block_run(self) -> None:
-        """Hold run() at its next resting point. Must be called on the loop."""
-        self._run_permit.clear()
-
-    def result(self, plan_return) -> RunEngineResult:
-        """Describe how the plan finished."""
-        return RunEngineResult(
-            tuple(self.run_start_uids),
-            plan_return,
-            self.exit_status,
-            self.interrupted,
-            self.reason,
-            self.exception,
-        )
-
-    @property
-    def state(self):
-        """The session's state machine. One session, one state."""
-        return self._session._state
-
-    @state.setter
-    def state(self, value):
-        self._session._state = value
-
-    def _default_commands(self):
-        """The vocabulary this executor understands out of the box."""
-        return {
-            "declare_stream": self._declare_stream,
-            "create": self._create,
-            "save": self._save,
-            "drop": self._drop,
-            "read": self._read,
-            "locate": self._locate,
-            "monitor": self._monitor,
-            "unmonitor": self._unmonitor,
-            "null": self._null,
-            "RE_class": self._RE_class,
-            "stop": self._stop,
-            "set": self._set,
-            "trigger": self._trigger,
-            "sleep": self._sleep,
-            "wait": self._wait,
-            "checkpoint": self._checkpoint,
-            "clear_checkpoint": self._clear_checkpoint,
-            "rewindable": self._rewindable,
-            "pause": self._pause,
-            "_resume_from_suspender": self._resume,
-            "_start_suspender": self._start_suspender,
-            "prepare": self._prepare,
-            "collect": self._collect,
-            "kickoff": self._kickoff,
-            "complete": self._complete,
-            "configure": self._configure,
-            "stage": self._stage,
-            "unstage": self._unstage,
-            "subscribe": self._subscribe,
-            "unsubscribe": self._unsubscribe,
-            "open_run": self._open_run,
-            "close_run": self._close_run,
-            "wait_for": self._wait_for,
-            "input": self._input,
-            "install_suspender": self._install_suspender,
-            "remove_suspender": self._remove_suspender,
-        }
-
-    def rebuild_command_registry(self):
-        """Compose the live vocabulary: built-ins, then the session's changes.
-
-        The session records what has been registered and unregistered, so that
-        those survive being composed over a different executor's built-ins.
-        """
-        registry = self._default_commands()
-        registry.update(self._session._registered_commands)
-        for name in self._session._unregistered_commands:
-            registry.pop(name, None)
-        self.command_registry = registry
-
-    def load_plan(self, plan, *, metadata=None, prologue=None):
-        """Put a plan on the stack, ready to be run.
-
-        Parameters
-        ----------
-        plan : iterable of Msg
-            The plan to execute. The session's preprocessors are applied to it.
-        metadata : dict, optional
-            Metadata for every run this plan opens.
-        prologue : iterable of Msg, optional
-            Messages to work off before the plan itself, used to wait for
-            tripped suspenders before starting.
-        """
-        self._plan = plan  # this ref is just used for metadata introspection
-        if metadata:
-            self._metadata_per_call.update(metadata)
-
-        gen = ensure_generator(plan)
-        for wrapper_func in self._session.preprocessors:
-            gen = wrapper_func(gen)
-
-        self._plan_stack.append(gen)
-        self._response_stack.append(None)
-        if prologue is not None:
-            self._plan_stack.append(prologue)
-            self._response_stack.append(None)
-
-    async def prepare_resume(self):
-        """Get ready to resume from a pause. Must be called on the loop."""
-        self.interrupted = False
-        for current_run in self._run_bundlers.values():
-            current_run.record_interruption("resume")
-        self._plan_stack.append(self.rewind())
-        self._response_stack.append(None)
-        # Notify Devices of the resume in case they want to clean up.
-        for obj in self._objs_seen:
-            if isinstance(obj, Pausable):
-                await maybe_await(obj.resume())
-
-    async def request_suspend(self, fut, *, pre_plan=None, post_plan=None, justification=None):
-        """Suspend until ``fut`` is finished. Must be called on the loop."""
-        if not self.resumable:
-            print("No checkpoint; cannot suspend.")
-            print("Aborting: running cleanup and marking exit_status as 'abort'...")
-            self.interrupted = True
-            self.exception = FailedPause()
-            was_paused = self.state == "paused"
-            self.state = "aborting"
-            if not was_paused:
-                self._task.cancel()
-        if justification is not None:
-            print(f"Justification for this suspension:\n{justification}")
-
-        # add starting the suspender logic to the stack
-        self._plan_stack.append(single_gen(Msg("_start_suspender", None, pre_plan, post_plan, justification, fut)))
-        self._response_stack.append(None)
-
-        # The event loop is still running. The pre_plan will be processed,
-        # and then the executor will be hung up on processing the
-        # 'wait_for' message until `fut` is set.
-        if not self.state == "paused":
-            self.state = "suspending"
-            # bump the run task out of what ever it is awaiting
-            self._task.cancel()
-
-    async def _stop_movable_objects(self, *, success=True):
-        "Call obj.stop() for all objects we have moved. Log any exceptions."
-        for obj in self._movable_objs_touched:
-            if isinstance(obj, Stoppable):
-                try:
-                    await maybe_await(obj.stop(success=success))
-                except Exception:
-                    self._session.log.exception("Failed to stop %r.", obj)
-            else:
-                self._session.log.debug("No 'stop' method available on %r", obj)
-
-    def _destroy_open_run_tracing_spans(self):
-        while len(self._run_tracing_spans):
-            _span = self._run_tracing_spans.pop()
-            _span.set_attribute("exit_status", "aborted")
-            _span.end()
-
-    async def run(self, plan=None, *, metadata=None):
-        """Execute ``plan``, or whatever has already been loaded.
-
-        Awaiting this is all that is needed to run a plan::
-
-            executor = PlanExecutor(session)
-            result = await executor.run(plan)
-
-        Parameters
-        ----------
-        plan : iterable of Msg, optional
-            The plan to execute. If omitted, whatever :meth:`load_plan` has
-            put on the stack is executed instead, which is how a `RunEngine`
-            uses this.
-        metadata : dict, optional
-            Metadata for every run the plan opens.
-
-        Returns
-        -------
-        The value the plan returned, or :data:`NO_PLAN_RETURN` if it did not
-        run to completion.
-
-        Notes
-        -----
-        Pull messages from the plan, process them, send results back.
-
-        Upon exit, clean up.
-        - Call stop() on all objects that were 'set' or 'kickoff'.
-        - Try to collect any uncollected flyers.
-        - Try to unstage any devices left staged by the plan.
-        - Try to remove any monitoring subscriptions left on by the plan.
-        - If interrupting the middle of a run, try to emit a RunStop document.
-        """
-        if plan is not None:
-            self.load_plan(plan, metadata=metadata)
-        await self._run_permit.wait()
-        # grab the current task.  We need to do this here because the
-        # object returned by `run_coroutine_threadsafe` is a future
-        # that acts as a proxy that does not have the correct behavior
-        # when `.cancel` is called on it.
-        self._task = asyncio.current_task(self._session.loop)
-        stashed_exception = None
-        debug = msg_logger.debug
-        self.reason = ""
-        # sentinel to decide if need to add to the response stack or not
-        sentinel = object()
-        plan_return = NO_PLAN_RETURN
-        exit_reason = ""
-        try:
-            self.state = "running"
-            while True:
-                if self.state in ("pausing", "suspending"):
-                    if not self.resumable:
-                        self._run_permit.set()
-                        stashed_exception = FailedPause()
-
-                        self.state = "aborting"
-                        continue
-                # currently only using 'suspending' to get us into the
-                # block above, we do not have a 'suspended' state
-                # (yet)
-                if self.state == "suspending":
-                    self.state = "running"
-                if not self._run_permit.is_set():
-                    # A pause has been requested. First, put everything in a
-                    # resting state.
-                    assert self.state == "pausing"
-                    # Remove any monitoring callbacks, but keep refs in
-                    # self._monitor_params to re-instate them later.
-                    for current_run in self._run_bundlers.values():
-                        await current_run.suspend_monitors()
-                    # During pause, all motors should be stopped. Call stop()
-                    # on every object we ever set().
-                    await self._stop_movable_objects(success=True)
-                    # Notify Devices of the pause in case they want to
-                    # clean up.
-                    for obj in self._objs_seen:
-                        if isinstance(obj, Pausable):
-                            try:
-                                await maybe_await(obj.pause())
-                            except NoReplayAllowed:
-                                self._reset_checkpoint_state_meth()
-                    self.state = "paused"
-                    # Let RunEngine.__call__ return...
-                    self._session._notify_paused()
-
-                    await self._run_permit.wait()
-                    # Restore any monitors
-                    for current_run in self._run_bundlers.values():
-                        await current_run.restore_monitors()
-                    if self.state == "paused":
-                        # may be called by 'resume', 'stop', 'abort', 'halt'
-                        self.state = "running"
-
-                    # If we are here, we have come back to life either to
-                    # continue (resume) or to clean up before exiting.
-
-                assert len(self._response_stack) == len(self._plan_stack)
-                # set resp to the sentinel so that if we fail in the sleep
-                # we do not add an extra response
-                resp = sentinel
-                try:
-                    # the new response to be added
-                    new_response = None
-
-                    # This 'await' must be here to ensure that this coroutine
-                    # breaks out of its current behavior before trying to get
-                    # the next message from the top of the generator stack in
-                    # case there has been a pause requested.  Without this the
-                    # next message after the pause may be processed first on
-                    # resume (instead of the first message in self._msg_cache).
-                    # This await also gives the co-routine for requesting
-                    # suspends a chance to run.
-
-                    # This sleep has to be inside of this try block so that any
-                    # of the 'async' exceptions get thrown in the correct
-                    # place.
-
-                    # If we are handling an exception, then burn through the
-                    # current plan stack before rather than allowing a pause or
-                    # suspension to try and finish firing.
-                    if stashed_exception is None:
-                        await asyncio.sleep(0)
-                    # always pop off a result, we are either sending it back in
-                    # or throwing an exception in, in either case the left hand
-                    # side of the yield in the plan will be moved past
-                    resp = self._response_stack.pop()
-                    # if any status tasks have failed, grab the exceptions.
-                    # give priority to things pushed in from outside
-                    if self.exception is not None:
-                        stashed_exception = self.exception
-                        self.exception = None
-                    # The case where we have a stashed exception
-                    if stashed_exception is not None or isinstance(resp, Exception):
-                        # throw the exception at the current plan
-                        try:
-                            msg = self._plan_stack[-1].throw(stashed_exception or resp)
-                        except Exception as e:
-                            # The current plan did not handle it,
-                            # maybe the next plan (if any) would like
-                            # to try
-                            self._plan_stack.pop()
-                            # we have killed the current plan, do not give
-                            # it a new response
-                            resp = sentinel
-                            # If there is at least one plan left in the stack,
-                            # stash the new exception go back to top
-                            if len(self._plan_stack):
-                                stashed_exception = e
-                                continue
-                            # no plans left and still an unhandled exception
-                            # re-raise to exit the infinite loop
-                            else:
-                                raise
-                        # clear the stashed exception, the top plan
-                        # handled it.
-                        else:
-                            stashed_exception = None
-                    # The normal case of clean operation
-                    else:
-                        try:
-                            msg = self._plan_stack[-1].send(resp)
-                        # We have exhausted the top generator
-                        except StopIteration:
-                            # pop the dead generator go back to the top
-                            self._plan_stack.pop()
-                            # we have killed the current plan, do not give
-                            # it a new response
-                            resp = sentinel
-                            if len(self._plan_stack):
-                                continue
-                            # or reraise to get out of the infinite loop
-                            else:
-                                raise
-                        # Any other exception that comes out of the plan
-                        except Exception as e:
-                            # pop the dead plan, stash the exception and
-                            # go to the top of the loop
-                            self._plan_stack.pop()
-                            # we have killed the current plan, do not give
-                            # it a new response
-                            resp = sentinel
-                            if len(self._plan_stack):
-                                stashed_exception = e
-                                continue
-                            # or reraise to get out of the infinite loop
-                            else:
-                                raise
-
-                    # if we have a message hook, call it
-                    if self._session.msg_hook is not None:
-                        self._session.msg_hook(msg)
-                    debug(
-                        "%s(%r, *%r **%r, run=%r)",
-                        msg.command,
-                        msg.obj,
-                        msg.args,
-                        msg.kwargs,
-                        msg.run,
-                        extra={"msg_command": msg.command},
-                    )
-
-                    # update the running set of all objects we have seen
-                    self._objs_seen.add(msg.obj)
-
-                    # if this message can be cached for rewinding, cache it
-                    if (
-                        self._msg_cache is not None
-                        and self.rewindable_flag
-                        and msg.command not in UNCACHEABLE_COMMANDS
-                    ):
-                        # We have a checkpoint.
-                        self._msg_cache.append(msg)
-
-                    # try to look up the coroutine to execute the command
-                    if (
-                        coro := self.command_registry.get(msg.command, key_absence_sentinel := object())
-                    ) is key_absence_sentinel:
-                        # flag invalid command
-                        # and return to the top of the loop
-                        new_response = InvalidCommand(msg.command)
-                        continue
-
-                    # try to finally run the command the user asked for
-                    try:
-                        # this is one of two places that 'async'
-                        # exceptions (coming in via throw) can be
-                        # raised
-                        new_response = await coro(msg)
-
-                    # special case `CancelledError` and let the outer
-                    # exception block deal with it.
-                    except asyncio.CancelledError:
-                        raise
-                    # any other exception, stash it and go to the top of loop
-                    except Exception as e:
-                        new_response = e
-                        continue
-                    # normal use, if it runs cleanly, stash the response and
-                    # go to the top of the loop
-                    else:
-                        continue
-
-                except KeyboardInterrupt:
-                    # This only happens if some external code captures SIGINT
-                    # -- overriding the RunEngine -- and then raises instead
-                    # of (properly) calling the RunEngine's handler.
-                    # See https://github.com/NSLS-II/bluesky/pull/242
-                    print(
-                        "An unknown external library has improperly raised "
-                        "KeyboardInterrupt. Intercepting and triggering "
-                        "a HALT."
-                    )
-                    await self.halt()
-                except asyncio.CancelledError as e:
-                    if self.state == "pausing":
-                        # if we got a CancelledError and we are in the
-                        # 'pausing' state clear the run permit and
-                        # bounce to the top
-                        self._run_permit.clear()
-                        continue
-                    if self.state in ("halting", "stopping", "aborting"):
-                        # if we got this while just keep going in tear-down
-                        exception_map = {"halting": PlanHalt, "stopping": RequestStop, "aborting": RequestAbort}
-                        # if the exception is not set bounce to the top
-                        if stashed_exception is None:
-                            stashed_exception = exception_map[self.state]
-                        continue
-                    if self.state == "suspending":
-                        # just bounce to the top
-                        continue
-                    # if we are handling this twice, raise and leave the plans
-                    # alone
-                    if stashed_exception is e:
-                        raise e
-                    # the case where FailedPause, RequestAbort or a coro
-                    # raised error is not already stashed in _exception
-                    if stashed_exception is None:
-                        stashed_exception = e
-                finally:
-                    # if we poped a response and did not pop a plan, we need
-                    # to put the new response back on the stack
-                    if resp is not sentinel:
-                        self._response_stack.append(new_response)
-
-        except StopIteration as e:
-            self.exit_status = "success"
-            plan_return = e.value
-            # TODO Is the sleep here necessary?
-            await asyncio.sleep(0)
-        except RequestStop:
-            self.exit_status = "success"
-            # TODO Is the sleep here necessary?
-            await asyncio.sleep(0)
-        except (FailedPause, RequestAbort, asyncio.CancelledError, PlanHalt):
-            self.exit_status = "abort"
-            # TODO Is the sleep here necessary?
-            await asyncio.sleep(0)
-            self._session.log.exception("Run aborted")
-        except GeneratorExit as err:
-            self.exit_status = "fail"  # Exception raises during 'running'
-            exit_reason = str(err)
-            raise ValueError from err
-        except Exception as err:
-            self.exit_status = "fail"  # Exception raises during 'running'
-            exit_reason = str(err)
-            self._session.log.exception("Run aborted")
-            raise err
-        finally:
-            if not exit_reason:
-                exit_reason = self.reason
-            # Some done_callbacks may still be alive in other threads.
-            # Block them from creating new 'failed status' tasks on the loop.
-            self._pardon_failures.set()
-            # call stop() on every movable object we ever set()
-            await self._stop_movable_objects(success=True)
-            for current_run in self._run_bundlers.values():
-                # Clear any uncleared monitoring callbacks.
-                current_run.clear_monitors()
-                # Try to collect any flyers that were kicked off but
-                # not finished.  Some might not support partial
-                # collection. We swallow errors.
-                await current_run.backstop_collect()
-            # in case we were interrupted between 'stage' and 'unstage'
-            for obj in list(self._staged):
-                try:
-                    obj.unstage()
-                except Exception:
-                    self._session.log.exception("Failed to unstage %r.", obj)
-                self._staged.remove(obj)
-
-            sys.stdout.flush()
-            # Emit RunStop if necessary.
-            for key, current_run in self._run_bundlers.items():
-                if current_run.run_is_open:
-                    try:
-                        await current_run.close_run(
-                            Msg("close_run", exit_status=self.exit_status, reason=exit_reason, run_id=key)
-                        )
-                    except Exception:
-                        self._session.log.error("Failed to close run %r.", current_run)
-            self._run_bundlers.clear()
-
-            for p in self._plan_stack:
-                try:
-                    p.close()
-                except RuntimeError:
-                    print(f"The plan {p!r} tried to yield a value on close.  Please fix your plan.")
-
-            self.state = "idle"
-
-        self._session.log.info("Cleaned up from plan %r", self._plan)
-        if isinstance(stashed_exception, asyncio.CancelledError):
-            raise stashed_exception
-        return plan_return
-
-    async def _wait_for(self, msg):
-        """Instruct the RunEngine to wait for futures and return the resulting tasks.
-
-        Expected message object is:
-
-            Msg('wait_for', None, awaitable_factories, **kwargs)
-
-        The keyword arguments will be passed through to `asyncio.wait`.
-
-        The callables in awaitable_factories must have the signature ::
-
-           def fut_fac() -> awaitable:
-               'This must work multiple times'
-
-        """
-
-        (futs,) = msg.args
-        futs = [asyncio.ensure_future(f()) for f in futs]
-        completed, pending = await asyncio.wait(futs, **msg.kwargs)
-        if pending:
-            raise WaitForTimeoutError("Plan failed to complete in the specified time")
-        return futs
-
-    async def _open_run(self, msg):
-        """Instruct the RunEngine to start a new "run"
-
-        Expected message object is:
-
-            Msg('open_run', None, **kwargs)
-
-        where **kwargs are any additional metadata that should go into
-        the RunStart document
-        """
-        _span = tracer.start_span(f"{_SPAN_NAME_PREFIX} run")
-        _set_span_msg_attributes(_span, msg)
-
-        self._run_tracing_spans.append(_span)
-
-        # TODO extract this from the Msg
-        run_key = msg.run
-        if run_key in self._run_bundlers:
-            raise IllegalMessageSequence("A 'close_run' message was not received before the 'open_run' message")
-
-        # Run scan_id calculation method
-        self._session.md["scan_id"] = await maybe_await(self._session.scan_id_source(self._session.md))
-
-        # For metadata below, info about plan passed to self.__call__ for.
-        plan_type = type(self._plan).__name__
-        plan_name = getattr(self._plan, "__name__", "")
-
-        # Combine metadata, in order of decreasing precedence:
-        md = ChainMap(
-            self._metadata_per_call,  # from kwargs to self.__call__
-            msg.kwargs,  # from 'open_run' Msg
-            {
-                "plan_type": plan_type,  # computed from self._plan
-                "plan_name": plan_name,
-            },
-            self._session.md,
-        )  # stateful, persistent metadata
-        # The metadata is final. Validate it now, at the last moment.
-        self._session.md_validator(dict(md))
-
-        # Apply normalizer at the same level of the validator
-        validated = self._session.md_normalizer(copy.deepcopy(md))
-
-        current_run = self._run_bundlers[run_key] = self._session.run_bundler_cls(
-            validated,
-            self._session.record_interruptions,
-            self._session.emit,
-            self._session.emit_sync,
-            self._session.log,
-            strict_pre_declare=self._session._require_stream_declaration,
-        )
-
-        new_uid = await current_run.open_run(msg)
-        self.run_start_uids.append(new_uid)
-        return new_uid
-
-    async def _close_run(self, msg):
-        """Instruct the RunEngine to write the RunStop document
-
-        Expected message object is:
-
-            Msg('close_run', None, exit_status=None, reason=None)
-
-        if *exit_stats* and *reason* are not provided, use the values
-        stashed on the RE.
-        """
-        # TODO extract this from the Msg
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object)
-        ) is key_absence_sentinel:
-            ims_msg = "A 'close_run' message was not received before the 'open_run' message"
-            raise IllegalMessageSequence(ims_msg)
-        ret = await current_run.close_run(msg)
-        del self._run_bundlers[run_key]
-        self._close_run_trace(msg)
-        return ret
-
-    def _close_run_trace(self, msg: Msg):
-        exit_status = msg.kwargs.get("exit_status", self.exit_status)
-        reason = msg.kwargs.get("reason", self.reason)
-        try:
-            _span: Span = self._run_tracing_spans.pop()
-            _span.set_attribute("exit_status", exit_status if exit_status is not None else "None")
-            _span.set_attribute("reason", reason if reason is not None else "None")
-            _span.end()
-        except IndexError:
-            logger.warning("No open traces left to close!")
-
-    async def _create(self, msg):
-        """Trigger the run engine to start bundling future obj.read() calls for
-         an Event document
-
-        Expected message object is:
-
-            Msg('create', None, name='primary')
-            Msg('create', name='primary')
-
-        Note that the `name` kwarg will be the 'name' field of the resulting
-        descriptor. So descriptor['name'] = msg.kwargs['name'].
-
-        Also note that changing the 'name' of the Event will create a new
-        Descriptor document.
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = (
-                "Cannot bundle readings without an open run. That is, 'create' must be preceded by 'open_run'."
-            )
-            raise IllegalMessageSequence(ims_msg)
-        return await current_run.create(msg)
-
-    async def _declare_stream(self, msg):
-        """Trigger the run engine to start bundling future obj.describe() calls for
-         an Event document
-
-        Expected message object is:
-
-            Msg('declare_stream', None, name='primary')
-            Msg('declare_stream', name='primary')
-            Msg('create', name='primary', collect=True)
-
-        Note that the `name` kwarg will be the 'name' field of the resulting
-        descriptor. So descriptor['name'] = msg.kwargs['name'].
-
-        If `collect` is set to True (default false) then `describe_collect` will be called
-        on declare_stream, rather than `describe`.
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = (
-                "Cannot bundle readings without an open run. That is, 'create' must be preceded by 'open_run'."
-            )
-            raise IllegalMessageSequence(ims_msg)
-        return await current_run.declare_stream(msg)
-
-    async def _read(self, msg):
-        """
-        Add a reading to the open event bundle.
-
-        Expected message object is:
-
-            Msg('read', obj)
-        """
-        obj = check_supports(msg.obj, Readable)
-        # actually _read_ the object
-        warn_if_msg_args_or_kwargs(msg, obj.read, msg.args, msg.kwargs)
-        ret = await maybe_await(obj.read(*msg.args, **msg.kwargs))
-
-        if ret is None:
-            raise RuntimeError(
-                f"The read of {obj.name} returned None. "
-                "This is a bug in your object implementation, "
-                "`read` must return a dictionary."
-            )
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is not key_absence_sentinel:
-            await current_run.read(msg, ret)
-
-        return ret
-
-    async def _locate(self, msg: Msg):
-        """
-        Locate some Movables and return their locations.
-
-        Expected message object is:
-
-            Msg('locate', obj1, ..., objn, squeeze=True)
-
-        If a single obj is passed, obj.locate() is returned. If multiple objs
-        are passed, obj.locate() is called in parallel for all objs and a list
-        of the results returned. If squeeze is supplied and is False then it
-        will always return a list of results even with a single object.
-        """
-        objs = [check_supports(obj, Locatable) for obj in (msg.obj,) + msg.args]
-        # actually _locate_ the objects
-        coros = [maybe_await(obj.locate()) for obj in objs]
-        if len(coros) == 1 and msg.kwargs.get("squeeze", True):
-            return await coros[0]
-        else:
-            return list(await asyncio.gather(*coros))
-
-    async def _monitor(self, msg):
-        """
-        Monitor a signal. Emit event documents asynchronously.
-
-        A descriptor document is emitted immediately. Then, a closure is
-        defined that emits Event documents associated with that descriptor
-        from a separate thread. This process is not related to the main
-        bundling process (create/read/save).
-
-        Expected message object is:
-
-            Msg('monitor', obj, **kwargs)
-            Msg('monitor', obj, name='event-stream-name', **kwargs)
-
-        where kwargs are passed through to ``obj.subscribe()``
-        """
-
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = "A 'monitor' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-        else:
-            await current_run.monitor(msg)
-        await self._reset_checkpoint_state_coro()
-
-    async def _unmonitor(self, msg):
-        """
-        Stop monitoring; i.e., remove the callback emitting event documents.
-
-        Expected message object is:
-
-            Msg('unmonitor', obj)
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = "An 'unmonitor' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-        else:
-            await current_run.unmonitor(msg)
-        await self._reset_checkpoint_state_coro()
-
-    async def _save(self, msg):
-        """Save the event that is currently being bundled
-
-        Expected message object is:
-
-            Msg('save')
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            # sanity check -- this should be caught by 'create' which makes
-            # this code path impossible
-            ims_msg = "A 'save' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-        else:
-            await current_run.save(msg)
-
-    async def _drop(self, msg):
-        """Drop the event that is currently being bundled
-
-        Expected message object is:
-
-            Msg('drop')
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = "A 'drop' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-        else:
-            await current_run.drop(msg)
-
-    async def _prepare(self, msg):
-        """Prepare a flyer for a flyscan
-
-        Expected message object is:
-
-        If `flyer_object` obeys the Preparable protocol, it should have a .prepare
-        method that takes an argument to be set:
-
-            Msg('prepare', flyer_object, value)
-
-        Where value represents an initial state to move the flyer to.
-        """
-        obj = check_supports(msg.obj, Preparable)
-        kwargs = dict(msg.kwargs)
-        group = kwargs.pop("group", None)
-        ret = obj.prepare(*msg.args, **kwargs)
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="prepare")
-
-        return ret
-
-    async def _kickoff(self, msg):
-        """Start a flyscan object
-
-        Special kwargs for the 'Msg' object in this function:
-        group : str
-            The blocking group to this flyer to
-
-        Expected message object is:
-
-        If `flyer_object` has a `kickoff` function that takes no arguments:
-
-            Msg('kickoff', flyer_object)
-            Msg('kickoff', flyer_object, group=<name>)
-
-        If `flyer_object` has a `kickoff` function that takes
-        `(start, stop, steps)` as its function arguments:
-
-            Msg('kickoff', flyer_object, start, stop, step)
-            Msg('kickoff', flyer_object, start, stop, step, group=<name>)
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            ims_msg = "A 'kickoff' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-
-        _, obj, args, kwargs, _ = msg
-        obj = check_supports(obj, Flyable)
-        kwargs = dict(msg.kwargs)
-        group = kwargs.pop("group", None)
-        warn_if_msg_args_or_kwargs(msg, obj.kickoff, msg.args, kwargs)
-        ret = obj.kickoff(*msg.args, **kwargs)
-        await current_run.kickoff(msg)
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="kickoff")
-
-        return ret
-
-    @tracer.start_as_current_span(f"{_SPAN_NAME_PREFIX} complete")
-    async def _complete(self, msg):
-        """
-        Tell a flyer, 'stop collecting, whenever you are ready'.
-
-        The flyer returns a status object. Some flyers respond to this
-        command by stopping collection and returning a finished status
-        object immediately. Other flyers finish their given course and
-        finish whenever they finish, irrespective of when this command is
-        issued.
-
-        Expected message object is:
-
-            Msg('complete', flyer, group=<GROUP>)
-
-        where <GROUP> is a hashable identifier.
-        """
-        _set_span_msg_attributes(trace.get_current_span(), msg)
-        kwargs = dict(msg.kwargs)
-        group = kwargs.pop("group", None)
-        obj = check_supports(msg.obj, Flyable)
-        warn_if_msg_args_or_kwargs(msg, obj.complete, msg.args, kwargs)
-        ret = obj.complete(*msg.args, **kwargs)
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="complete")
-
-        return ret
-
-    @tracer.start_as_current_span(f"{_SPAN_NAME_PREFIX} collect")
-    async def _collect(self, msg):
-        """
-        Collect data cached by a flyer and emit documents
-
-        Expected message object is:
-
-            Msg('collect', flyer_object)
-            Msg('collect', flyer_object, stream=True, return_payload=False, name="a_name")
-        """
-        _set_span_msg_attributes(trace.get_current_span(), msg)
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            # TODO add test exercising this path
-            ims_msg = "A 'collect' message was sent but no run is open."
-            raise IllegalMessageSequence(ims_msg)
-
-        return await current_run.collect(msg)
-
-    async def _null(self, msg):
-        """
-        A no-op message, mainly for debugging and testing.
-        """
-        pass
-
-    async def _RE_class(self, msg):
-        """
-        A no-op message, mainly for debugging and testing.
-        """
-        return self._session.run_engine_cls or type(self)
-
-    @tracer.start_as_current_span(f"{_SPAN_NAME_PREFIX} set")
-    async def _set(self, msg):
-        """
-        Set a device and cache the returned status object.
-
-        Also, note that the device has been touched so it can be stopped upon
-        exit.
-
-        Expected message object is
-
-            Msg('set', obj, *args, **kwargs)
-
-        where arguments are passed through to `obj.set(*args, **kwargs)`.
-        """
-        _set_span_msg_attributes(trace.get_current_span(), msg)
-        obj = check_supports(msg.obj, Movable)
-        kwargs = dict(msg.kwargs)
-        group = kwargs.pop("group", None)
-        self._movable_objs_touched.add(obj)
-        ret = obj.set(*msg.args, **kwargs)
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="set")
-
-        return ret
-
-    async def _trigger(self, msg):
-        """
-        Trigger a device and cache the returned status object.
-
-        Expected message object is:
-
-            Msg('trigger', obj)
-        """
-        obj = check_supports(msg.obj, Triggerable)
-        kwargs = dict(msg.kwargs)
-        group = kwargs.pop("group", None)
-        warn_if_msg_args_or_kwargs(msg, obj.trigger, msg.args, kwargs)
-        ret = obj.trigger(*msg.args, **kwargs)
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="trigger")
-
-        return ret
-
-    @tracer.start_as_current_span(f"{_SPAN_NAME_PREFIX} wait")
-    async def _wait(self, msg: Msg) -> bool:
-        """Block progress until every object that was triggered or set
-        with the keyword argument `group=<GROUP>` is done. Returns a boolean that is
-        true when all triggered objects are done. When the keyword argument
-        `error_on_timeout=<error_on_timeout>` is false, this method can return before all objects are done
-        after a flush period given by the `timeout=<TIMEOUT>` keyword argument.
-
-        Expected message object is:
-
-            Msg('wait', group=<GROUP>, error_on_timeout=<ERROR_ON_TIMEOUT>)
-
-        where ``<GROUP>`` is any hashable key and ``<ERROR_ON_TIMEOUT>`` is a boolean.
-        """
-        _set_span_msg_attributes(trace.get_current_span(), msg)
-        done = False  # boolean that tracks whether waiting is complete
-        if msg.args:
-            (group,) = msg.args
-        else:
-            group = msg.kwargs["group"]
-        error_on_timeout = msg.kwargs.get("error_on_timeout", True)
-        watch = msg.kwargs.get("watch", ())
-        watch_task: asyncio.Task | None = None
-        if group:
-            trace.get_current_span().set_attribute("group", group)
-        else:
-            trace.get_current_span().set_attribute("no_group_given", True)
-        futs = self._groups.pop(group, set())
-        if futs:
-            status_objs = self._status_objs.pop(group)
-            try:
-                if not error_on_timeout:
-                    if group not in self._seen_wait_and_move_on_keys:
-                        self._seen_wait_and_move_on_keys.add(group)
-                        self._session._call_waiting_hook(status_objs)
-                else:  # if error_on_timeout False
-                    # Notify the waiting_hook function that the RunEngine is
-                    # waiting for these status_objs to complete. Users can use
-                    # the information these encapsulate to create a progress
-                    # bar.
-                    self._session._call_waiting_hook(status_objs)
-
-                async def wait_for_first_exception(futures: set) -> list[asyncio.Future]:
-                    return await self._wait_for(
-                        Msg(
-                            "wait_for",
-                            None,
-                            futures,
-                            return_when=asyncio.FIRST_EXCEPTION,
-                            timeout=msg.kwargs.get("timeout", None),
-                        )
-                    )
-
-                # Create the task waiting for the given group of statuses to complete
-                # or one of them to fail
-                status_task = asyncio.create_task(wait_for_first_exception(futs))
-                if watch:
-                    # Create a task that waits for an exception on any watch group
-                    # so we know whether to stop the wait early because of a watcher failure
-                    watch_futs = set()
-                    for w in watch:
-                        watch_futs.update(self._groups.get(w, set()))
-                    watch_task = asyncio.create_task(wait_for_first_exception(watch_futs))
-
-                    def cancel_status_task_if_error(fut: asyncio.Future[list[asyncio.Future]]):
-                        # If _wait_for raised an exception, or if any of the status
-                        # objects in the watch groups failed, cancel the status_task.
-                        if fut.exception() or any(f.exception() for f in fut.result()):
-                            status_task.cancel()
-
-                    watch_task.add_done_callback(cancel_status_task_if_error)
-                await status_task
-            except WaitForTimeoutError:
-                # We might wait to call wait again, so put the futures and status objects back in
-                self._groups[group] = futs
-                self._status_objs[group] = status_objs
-                if error_on_timeout:
-                    raise
-            finally:
-                if watch_task:
-                    watch_task.cancel()
-                if error_on_timeout:
-                    # Notify the waiting_hook function that we have moved on by
-                    # sending it `None`. If all goes well, it could have
-                    # inferred this from the status_obj, but there are edge
-                    # cases.
-                    self._session._call_waiting_hook(None)
-                    done = True
-                else:
-                    done = all(obj.done for obj in status_objs)
-                    if done:
-                        self._session._call_waiting_hook(None)
-                        self._seen_wait_and_move_on_keys.remove(group)
-        else:
-            done = True
-        return done
-
-    def _status_object_completed(self, ret, fut: asyncio.Future, pardon_failures, obj=None, action=None):
-        """
-        Task to run when a status object is finished.
-
-        Always called on the event loop, via the trampoline that
-        :meth:`_add_status_to_group` hands to the status object.
-
-        Parameters
-        ----------
-        ret : status object
-        p_event : asyncio.Future
-            held in the RunEngine's self._groups cache for waiting
-        pardon_failuers : asyncio.Event
-            tells us whether the __call__ this status object is over
-        obj : object, optional
-            the device the status object came from, for logging
-        action : str, optional
-            what the device was asked to do, for logging
-        """
-        self._session.log.debug("The object %r reports %r is done with status %r.", obj, action, ret.success)
-        if not ret.success and not pardon_failures.is_set():
-            # TODO: need a better channel to move this information back
-            # to the run task.
-            try:
-                exc = ret.exception(timeout=0)
-                raise FailedStatus(ret) from exc
-            except Exception as e:
-                self.exception = e
-                fut.set_exception(e)
-                # We have set the exception, but we don't mind if
-                # no-one collects it from the future, so fetch it ourselves to
-                # squash "Future exception was never retrieved" at teardown.
-                fut.exception()
-        else:
-            fut.set_result(None)
-
-    async def _sleep(self, msg):
-        """
-        Sleep the event loop.
-
-        Expected message object is:
-
-            Msg('sleep', None, sleep_time)
-
-        where `sleep_time` is in seconds
-        """
-        await asyncio.sleep(*msg.args)
-
-    async def _pause(self, msg):
-        """Request the run engine to pause
-
-        Expected message object is:
-
-            Msg('pause', defer=False, name=None, callback=None)
-
-        See RunEngine.request_pause() docstring for explanation of the three
-        keyword arguments in the `Msg` signature
-        """
-        await self.request_pause(*msg.args, **msg.kwargs)
-
-    async def _resume(self, msg):
-        """Request the run engine to resume
-
-        Expected message object is:
-
-            Msg('resume', defer=False, name=None, callback=None)
-
-        See RunEngine.resume() docstring for explanation of the three
-        keyword arguments in the `Msg` signature
-        """
-        # Re-instate monitoring callbacks.
-        for current_run in self._run_bundlers.values():
-            await current_run.restore_monitors()
-        # Notify Devices of the resume in case they want to clean up.
-        for obj in self._objs_seen:
-            if isinstance(obj, Pausable):
-                await maybe_await(obj.resume())
-
-    async def _checkpoint(self, msg):
-        """Instruct the RunEngine to create a checkpoint so that we can rewind
-        to this point if necessary
-
-        Expected message object is:
-
-            Msg('checkpoint')
-        """
-        for current_run in self._run_bundlers.values():
-            if current_run.bundling:
-                raise IllegalMessageSequence("Cannot 'checkpoint' after 'create' and before 'save'. Aborting!")
-
-        await self._reset_checkpoint_state_coro()
-
-        if self._deferred_pause_requested:
-            # We are at a checkpoint; we are done deferring the pause.
-            # Give the _check_for_signals coroutine time to look for
-            # additional SIGINTs that would trigger an abort.
-            await asyncio.sleep(0.5)
-            await self.request_pause(defer=False)
-
-    def _reset_checkpoint_state(self):
-        self._reset_checkpoint_state_meth()
-
-    def _reset_checkpoint_state_meth(self):
-        if self._msg_cache is None:
-            return
-
-        self._msg_cache = deque()
-        for current_run in self._run_bundlers.values():
-            current_run.reset_checkpoint_state()
-
-    async def _reset_checkpoint_state_coro(self):
-        self._reset_checkpoint_state()
-
-    async def _clear_checkpoint(self, msg):
-        """Clear a set checkpoint
-
-        Expected message object is:
-
-            Msg('clear_checkpoint')
-        """
-        # clear message cache
-        self._msg_cache = None
-        # clear stashed
-        for current_run in self._run_bundlers.values():
-            await current_run.clear_checkpoint(msg)
-
-    async def _rewindable(self, msg):
-        """Set rewindable state of RunEngine
-
-        Expected message object is:
-
-            Msg('rewindable', None, bool or None)
-        """
-
-        (rw_flag,) = msg.args
-        if rw_flag is not None:
-            self.rewindable_flag = rw_flag
-
-        return self.rewindable_flag
-
-    async def _configure(self, msg):
-        """Configure an object
-
-        Expected message object is:
-
-            Msg('configure', object, *args, **kwargs)
-
-        which results in this call:
-
-            object.configure(*args, **kwargs)
-        """
-        run_key = msg.run
-        if (
-            current_run := self._run_bundlers.get(run_key, key_absence_sentinel := object())
-        ) is key_absence_sentinel:
-            current_run = None
-        elif current_run.bundling:
-            ims_msg = "Cannot configure after 'create' but before 'save' Aborting!"
-            raise IllegalMessageSequence(ims_msg)
-        _, obj, args, kwargs, _ = msg
-
-        old, new = obj.configure(*args, **kwargs)
-        if current_run:
-            await current_run.configure(msg)
-        return old, new
-
-    def _add_status_to_group(self, obj: typing.Any, status_object: Status, group: str, action: str) -> None:
-        loop = self._session.loop
-        fut = loop.create_future()
-        pardon_failures = self._pardon_failures
-        settle = functools.partial(self._status_object_completed, status_object, fut, pardon_failures, obj, action)
-
-        # A sync ophyd Status runs its callbacks on whichever thread completed
-        # it, so this may be called from a thread that is not the loop's. It
-        # does nothing but hop back onto the loop, which keeps
-        # _status_object_completed, and so all of the state it touches, on the
-        # loop thread. Any arguments the device passes are dropped: settle has
-        # closed over what it needs. An ophyd-async status already calls back
-        # on the loop, where call_soon_threadsafe remains correct.
-        def done_callback(*args: typing.Any, **kwargs: typing.Any) -> None:
-            loop.call_soon_threadsafe(settle)
-
-        try:
-            status_object.add_callback(done_callback)
-        except AttributeError:
-            # for ophyd < v0.8.0
-            status_object.finished_cb = done_callback  # type: ignore
-
-        self._groups[group].add(lambda: fut)
-        self._status_objs[group].add(status_object)
-
-    async def _stage(self, msg):
-        """Instruct the RunEngine to stage the object
-
-        Expected message object is:
-
-            Msg('stage', object)
-        """
-        _, obj, args, kwargs, _ = msg
-        # If an object has no 'stage' method, assume there is nothing to do.
-        if not isinstance(obj, Stageable):
-            return []
-        group = kwargs.pop("group", None)
-        ret = obj.stage()
-        self._staged.add(obj)  # add first in case of failure below
-        await self._reset_checkpoint_state_coro()
-
-        if not isinstance(ret, Status):
-            return ret
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="stage")
-
-        return ret
-
-    async def _unstage(self, msg):
-        """Instruct the RunEngine to unstage the object
-
-        Expected message object is:
-
-            Msg('unstage', object)
-        """
-        _, obj, args, kwargs, _ = msg
-        # If an object has no 'unstage' method, assume there is nothing to do.
-        if not isinstance(obj, Stageable):
-            return []
-        group = kwargs.pop("group", None)
-        ret = obj.unstage()
-        # use `discard()` to ignore objects that are not in the staged set.
-        self._staged.discard(obj)
-        await self._reset_checkpoint_state_coro()
-
-        if not isinstance(ret, Status):
-            return ret
-
-        self._add_status_to_group(obj=obj, status_object=ret, group=group, action="unstage")
-
-        return ret
-
-    async def _stop(self, msg):
-        """
-        Stop a device.
-
-        Expected message object is:
-
-            Msg('stop', obj)
-        """
-        obj = check_supports(msg.obj, Stoppable)
-        return await maybe_await(obj.stop())  # nominally, this returns None
-
-    async def _subscribe(self, msg):
-        """
-        Add a subscription after the run has started.
-
-        This, like subscriptions passed to __call__, will be removed at the
-        end by the RunEngine.
-
-        Expected message object is:
-
-            Msg('subscribe', None, callback_function, document_name)
-
-        where `document_name` is one of:
-
-            {'start', 'descriptor', 'event', 'stop', 'all'}
-
-        and `callback_function` is expected to have a signature of:
-
-            ``f(name, document)``
-
-            where name is one of the ``document_name`` options and ``document``
-            is one of the document dictionaries in the event model.
-
-        See the docstring of bluesky.run_engine.Dispatcher.subscribe() for more
-        information.
-        """
-        self._session.log.debug("Adding subscription %r", msg)
-        _, obj, args, kwargs, _ = msg
-        token = self._session.subscribe(*args, **kwargs)
-        self.add_temp_subscription(token)
-        await self._reset_checkpoint_state_coro()
-        return token
-
-    async def _unsubscribe(self, msg):
-        """
-        Remove a subscription during a call -- useful for a multi-run call
-        where subscriptions are wanted for some runs but not others.
-
-        Expected message object is:
-
-            Msg('unsubscribe', None, TOKEN)
-            Msg('unsubscribe', token=TOKEN)
-
-        where ``TOKEN`` is the return value from ``RunEngine._subscribe()``
-        """
-        self._session.log.debug("Removing subscription %r", msg)
-        _, obj, arg, kwargs, _ = msg
-        if (token := kwargs.get("token", key_absence_sentinel := object())) is key_absence_sentinel:
-            (token,) = arg
-        self.drop_temp_subscription(token)
-        await self._reset_checkpoint_state_coro()
-
-    async def _input(self, msg):
-        """
-        Process a 'input' Msg. Expected Msg:
-
-            Msg('input', None)
-            Msg('input', None, prompt='>')  # customize prompt
-        """
-        prompt = msg.kwargs.get("prompt", "")
-        async_input = AsyncInput(self._session.loop)
-        async_input = functools.partial(async_input, end="", flush=True)
-        return await async_input(prompt)
-
-    def rewind(self):
-        """Clean up in preparation for resuming from a pause or suspension.
-
-        Returns
-        -------
-        new_plan : generator
-             A new plan made from the messages in the message cache
-
-        """
-        len_msg_cache = len(self._msg_cache)
-        new_plan = ensure_generator(list(self._msg_cache))
-        self._msg_cache = deque()
-        if len_msg_cache:
-            for current_run in self._run_bundlers.values():
-                current_run.rewind()
-
-        return new_plan
-
-    async def _start_suspender(self, msg):
-        """
-        An internal message to do the initial work of starting a suspender
-        """
-        pre_plan, post_plan, justification, fut = msg.args
-        for current_run in self._run_bundlers.values():
-            current_run.record_interruption(justification if justification is not None else "suspended")
-        # During suspend, all motors should be stopped. Call stop() on
-        # every object we ever set().
-        await self._stop_movable_objects(success=True)
-        # Notify Devices of the pause in case they want to clean up.
-        for obj in self._objs_seen:
-            if hasattr(obj, "pause"):
-                try:
-                    await maybe_await(obj.pause())
-                except NoReplayAllowed:
-                    self._reset_checkpoint_state_meth()
-        # rewind to the last checkpoint
-        rewind_plan = self.rewind()
-        was_rewindable = self.rewindable_flag
-
-        if callable(pre_plan):
-            pre_plan = pre_plan()
-        if callable(post_plan):
-            post_plan = post_plan()
-
-        def suspender_helper_inner_plan():
-            # none of this should run again.
-            yield Msg("rewindable", None, False)
-            # if there is a pre plan add on top of the wait
-            if pre_plan is not None:
-                yield from ensure_generator(pre_plan)
-            # wait for the future from the suspender to be released
-            yield Msg(
-                "wait_for",
-                None,
-                [
-                    fut,
-                ],
-            )
-            # do the work we need to do to resume
-            yield Msg(
-                "_resume_from_suspender",
-                None,
-            )
-            # if there is a post plan, run it
-            if post_plan is not None:
-                yield from ensure_generator(post_plan)
-            # put rewindable back the way it was
-            yield Msg("rewindable", None, was_rewindable)
-            yield from rewind_plan
-
-        # add the above helper to the plan stack
-        self._plan_stack.append(suspender_helper_inner_plan())
-        self._response_stack.append(None)
-
-    async def _install_suspender(self, msg):
-        """
-        See :meth: `RunEngine.install_suspender`
-
-        Expected message object is:
-
-            Msg('install_suspender', None, suspender)
-        """
-        suspender = msg.args[0]
-        self._session.install_suspender(suspender)
-
-    async def _remove_suspender(self, msg):
-        """
-        See :meth: `RunEngine.remove_suspender`
-
-        Expected message object is:
-
-            Msg('remove_suspender', None, suspender)
-        """
-        suspender = msg.args[0]
-        self._session.remove_suspender(suspender)
-
-    async def request_pause(self, defer=False):
-        # We are pausing. Cancel any deferred pause previously requested.
-        if not self.state.can_pause:
-            raise TransitionError(f"Run Engine is in '{self.state}' state and can not be paused.")
-
-        if defer:
-            self._deferred_pause_requested = True
-            print("Deferred pause acknowledged. Continuing to checkpoint.")
-            return
-
-        print("Pausing...")
-
-        self._deferred_pause_requested = False
-        self.interrupted = True
-        self.state = "pausing"
-        for current_run in self._run_bundlers.values():
-            current_run.record_interruption("pause")
-
-        self._task.cancel()
-
-    async def abort(self, reason=""):
-        if self.state.is_idle:
-            raise TransitionError("RunEngine is already idle.")
-        print("Aborting: running cleanup and marking exit_status as 'abort'...")
-        self.interrupted = True
-        self.reason = reason
-
-        self.exit_status = "abort"
-        self._destroy_open_run_tracing_spans()
-
-        was_paused = self.state == "paused"
-        self.state = "aborting"
-        if was_paused:
-            self.exception = RequestAbort()
-        else:
-            self._task.cancel()
-
-    async def stop(self):
-        if self.state.is_idle:
-            raise TransitionError("RunEngine is already idle.")
-        print("Stopping: running cleanup and marking exit_status as 'success'...")
-
-        self.interrupted = True
-        was_paused = self.state == "paused"
-        self.state = "stopping"
-        if was_paused:
-            self.exception = RequestStop
-        else:
-            self._task.cancel()
-
-    async def halt(self):
-        if self.state.is_idle:
-            raise TransitionError("RunEngine is already idle.")
-        print("Halting: skipping cleanup and marking exit_status as 'abort'...")
-        self._destroy_open_run_tracing_spans()
-        self.interrupted = True
-        was_paused = self.state == "paused"
-        self.state = "halting"
-        if was_paused:
-            self.exception = PlanHalt
-            self.exit_status = "abort"
-        else:
-            self._task.cancel()
+PAUSE_MSG = """
+Your RunEngine is entering a paused state. These are your options for changing
+the state of the RunEngine:
+
+RE.resume()    Resume the plan.
+RE.abort()     Perform cleanup, then kill plan. Mark exit_stats='aborted'.
+RE.stop()      Perform cleanup, then kill plan. Mark exit_status='success'.
+RE.halt()      Emergency Stop: Do not perform cleanup --- just stop.
+"""
+
+
+MAX_DEPTH_EXCEEDED_ERR_MSG = """
+RunEngine.max_depth is set to {}; depth of {} was detected.
+
+The RunEngine should not be called from inside another function. Doing so
+breaks introspection tools and can result in unexpected behavior in the event
+of an interruption. See documentation for more information and what to do
+instead:
+
+http://nsls-ii.github.io/bluesky/plans_intro.html#combining-plans
+"""
 
 
 class RunEngine:
@@ -2521,6 +464,10 @@ class RunEngine:
         return self._session._suspenders
 
     @property
+    def _temp_callback_ids(self):
+        return self._executor._temp_callback_ids
+
+    @property
     def commands(self):
         """
         The list of commands available to Msg.
@@ -2720,9 +667,7 @@ class RunEngine:
         :meth:`RunEngine.print_command_registry`
         :attr:`RunEngine.commands`
         """
-        self._session._registered_commands[name] = func
-        self._session._unregistered_commands.discard(name)
-        self._rebuild_command_registry()
+        self._session.register_command(name, func)
 
     def unregister_command(self, name):
         """
@@ -2738,12 +683,7 @@ class RunEngine:
         :meth:`RunEngine.print_command_registry`
         :attr:`RunEngine.commands`
         """
-        # Raise KeyError for an unknown command, as deleting from the registry
-        # directly used to.
-        del self._executor.command_registry[name]
-        self._session._registered_commands.pop(name, None)
-        self._session._unregistered_commands.add(name)
-        self._rebuild_command_registry()
+        self._session.unregister_command(name)
 
     async def _request_pause_coro(self, defer=False):
         """Pause without blocking the caller.
@@ -2877,7 +817,7 @@ class RunEngine:
 
         def _build_task():
             # make sure _run will block at the top
-            self._executor._run_permit.clear()
+            self._executor.block_run()
             self._blocking_event.clear()
             self._task_fut = asyncio.run_coroutine_threadsafe(self._executor.run(), loop=self.loop)
 
@@ -2948,7 +888,7 @@ class RunEngine:
                 except concurrent.futures.CancelledError:
                     return NO_PLAN_RETURN
             # The _run task is waiting on this Event. Let is continue.
-            self.loop.call_soon_threadsafe(self._executor._run_permit.set)
+            self.loop.call_soon_threadsafe(self._executor.permit_run)
             try:
                 # Block until plan is complete or exception is raised.
                 try:
@@ -3165,22 +1105,15 @@ class RunEngine:
         was_paused = self._state == "paused"
         self.loop.call_soon_threadsafe(start_task)
         coro_event.wait()
-        # Surface anything the interrupt coroutine raised, then take the
-        # result before resuming rather than after. Resuming runs the plan's
-        # cleanup, which opens and closes runs of its own, so a result read
-        # afterwards describes the cleanup as much as the plan.
+        # Re-raise anything the coroutine raised, e.g. a TransitionError.
         task.result()
+        # Describe the outcome before resuming, since resuming lets the plan
+        # run its cleanup and change what we would report.
         result = self._interrupted_result()
         if was_paused:
             self._resume_task()
 
         return result
-
-    def _interrupted_result(self):
-        """What abort(), stop() and halt() return."""
-        if self._call_returns_result:
-            return self._create_result(NO_PLAN_RETURN)
-        return tuple(self._run_start_uids)
 
     def _interrupted_result(self):
         """What abort(), stop() and halt() return."""
@@ -3198,233 +1131,6 @@ class RunEngine:
         self.emit_sync(name, doc)
 
 
-class Dispatcher:
-    """Dispatch documents to user-defined consumers on the main thread."""
-
-    def __init__(self):
-        self.cb_registry = CallbackRegistry(allowed_sigs=DocumentNames)
-        self._counter = count()
-        self._token_mapping = dict()  # noqa: C408
-
-    def process(self, name, doc):
-        """
-        Dispatch document ``doc`` of type ``name`` to the callback registry.
-
-        Parameters
-        ----------
-        name : {'start', 'descriptor', 'event', 'stop'}
-        doc : dict
-        """
-        exceptions = self.cb_registry.process(name, name.name, doc)
-        for exc, traceback in exceptions:  # noqa: B007
-            warn(  # noqa: B028
-                "A %r was raised during the processing of a %s "  # noqa: UP031
-                "Document. The error will be ignored to avoid "
-                "interrupting data collection. To investigate, "
-                "set RunEngine.ignore_callback_exceptions = False "
-                "and run again." % (exc, name.name)
-            )
-
-    def subscribe(self, func, name="all"):
-        """
-        Register a callback function to consume documents.
-
-        .. versionchanged :: 0.10.0
-            The order of the arguments was swapped and the ``name``
-            argument has been given a default value, ``'all'``. Because the
-            meaning of the arguments is unambiguous (they must be a callable
-            and a string, respectively) the old order will be supported
-            indefinitely, with a warning.
-
-        .. versionchanged :: 0.10.0
-            The order of the arguments was swapped and the ``name``
-            argument has been given a default value, ``'all'``. Because the
-            meaning of the arguments is unambiguous (they must be a callable
-            and a string, respectively) the old order will be supported
-            indefinitely, with a warning.
-
-        Parameters
-        ----------
-        func: callable
-            expecting signature like ``f(name, document)``
-            where name is a string and document is a dict
-        name : {'all', 'start', 'descriptor', 'event', 'stop'}, optional
-            the type of document this function should receive ('all' by
-            default).
-
-        Returns
-        -------
-        token : int
-            an integer ID that can be used to unsubscribe
-
-        See Also
-        --------
-        :meth:`Dispatcher.unsubscribe`
-            an integer token that can be used to unsubscribe
-        """
-        if callable(name) and isinstance(func, str):
-            name, func = func, name
-            warn(  # noqa: B028
-                "The order of the arguments has been changed. Because the "
-                "meaning of the arguments is unambiguous, the old usage will "
-                "continue to work indefinitely, but the new usage is "
-                "encouraged: call subscribe(func, name) instead of "
-                "subscribe(name, func). Additionally, the 'name' argument "
-                "has become optional. Its default value is 'all'."
-            )
-        if name == "all":
-            private_tokens = []
-            for key in DocumentNames:
-                private_tokens.append(self.cb_registry.connect(key, func))
-            public_token = next(self._counter)
-            self._token_mapping[public_token] = private_tokens
-            return public_token
-
-        name = DocumentNames[name]
-        private_token = self.cb_registry.connect(name, func)
-        public_token = next(self._counter)
-        self._token_mapping[public_token] = [private_token]
-        return public_token
-
-    def unsubscribe(self, token):
-        """
-        Unregister a callback function using its integer ID.
-
-        Parameters
-        ----------
-        token : int
-            the integer ID issued by :meth:`Dispatcher.subscribe`
-
-        See Also
-        --------
-        :meth:`Dispatcher.subscribe`
-        """
-        for private_token in self._token_mapping.pop(token, []):
-            self.cb_registry.disconnect(private_token)
-
-    def unsubscribe_all(self):
-        """Unregister all callbacks from the dispatcher."""
-        for public_token in list(self._token_mapping.keys()):
-            self.unsubscribe(public_token)
-
-    @property
-    def ignore_exceptions(self):
-        return self.cb_registry.ignore_exceptions
-
-    @ignore_exceptions.setter
-    def ignore_exceptions(self, val):
-        self.cb_registry.ignore_exceptions = val
-
-
-PAUSE_MSG = """
-Your RunEngine is entering a paused state. These are your options for changing
-the state of the RunEngine:
-
-RE.resume()    Resume the plan.
-RE.abort()     Perform cleanup, then kill plan. Mark exit_stats='aborted'.
-RE.stop()      Perform cleanup, then kill plan. Mark exit_status='success'.
-RE.halt()      Emergency Stop: Do not perform cleanup --- just stop.
-"""
-
-
-MAX_DEPTH_EXCEEDED_ERR_MSG = """
-RunEngine.max_depth is set to {}; depth of {} was detected.
-
-The RunEngine should not be called from inside another function. Doing so
-breaks introspection tools and can result in unexpected behavior in the event
-of an interruption. See documentation for more information and what to do
-instead:
-
-http://nsls-ii.github.io/bluesky/plans_intro.html#combining-plans
-"""
-
-
-def _set_span_msg_attributes(span, msg):
-    span.set_attribute("msg.command", msg.command)
-    span.set_attribute("msg.args", sanitize_np(msg.args))
-    span.set_attribute("msg.kwargs", json.dumps(msg.kwargs, default=repr))
-    span.set_attribute("msg.obj", repr(msg.obj)) if msg.obj else span.set_attribute("msg.no_obj_given", True)
-
-
-def _default_md_validator(md: RunEngineMetadata) -> None:
-    if "sample" in md and not (hasattr(md["sample"], "keys") or isinstance(md["sample"], str)):
-        raise ValueError(
-            "You specified 'sample' metadata. We give this field special "
-            "significance in order to make your data easily searchable. "
-            "Therefore, you must make 'sample' a string or a  "
-            "dictionary, like so: "
-            "GOOD: sample='dirt' "
-            "GOOD: sample={'color': 'red', 'number': 5} "
-            "BAD: sample=[1, 2] "
-        )
-
-
-def _default_md_normalizer(md: RunEngineMetadata) -> RunEngineMetadata:
-    return md
-
-
-def _ensure_event_loop_running(loop):
-    """
-    Run an asyncio event loop forever on a background thread.
-
-    This is idempotent: if the loop is already running nothing will be done.
-    """
-    if not loop.is_running():
-        th = threading.Thread(target=loop.run_forever, daemon=True, name="bluesky-run-engine")
-        th.start()
-        _ensure_event_loop_running.loop_to_thread[loop] = th
-    else:
-        th = _ensure_event_loop_running.loop_to_thread[loop]
-    return th
-
-
-_ensure_event_loop_running.loop_to_thread = weakref.WeakKeyDictionary()  # type: ignore
-
-_bluesky_event_loop = None
-
-
-def get_bluesky_event_loop():
-    return _bluesky_event_loop
-
-
-def set_bluesky_event_loop(loop):
-    global _bluesky_event_loop
-    _bluesky_event_loop = loop
-
-
-def in_bluesky_event_loop() -> bool:
-    try:
-        loop = asyncio.get_running_loop()
-    except RuntimeError:
-        # Ok, no running loop
-        return False
-    else:
-        # Check if running loop is bluesky event loop
-        return loop is _bluesky_event_loop
-
-
-def call_in_bluesky_event_loop(coro: typing.Awaitable[T], timeout: float | None = None) -> T:
-    if _bluesky_event_loop is None or not _bluesky_event_loop.is_running():
-        # Quell "coroutine never awaited" warnings
-        if iscoroutine(coro):
-            coro.close()
-        raise RuntimeError("Bluesky event loop not running")
-    fut: concurrent.futures.Future = asyncio.run_coroutine_threadsafe(
-        coro,  # type: ignore
-        loop=_bluesky_event_loop,
-    )
-    return fut.result(timeout=timeout)
-
-
-def autoawait_in_bluesky_event_loop(ip=None):
-    if ip is None:
-        import IPython
-
-        ip = IPython.get_ipython()  # type: ignore
-    assert ip, "Couldn't import IPython"
-    ip.loop_runner = call_in_bluesky_event_loop
-
-
 # Names the RunEngine used to hold itself, which now belong to the executor
 # for the plan being run. Mapped to the executor attribute they forward to.
 #
@@ -3435,7 +1141,6 @@ def autoawait_in_bluesky_event_loop(ip=None):
 # moved to reading RunEngine._executor, or to using a PlanExecutor directly.
 _EXECUTOR_FORWARDS = {
     "_run_permit": "_run_permit",
-    "_temp_callback_ids": "_temp_callback_ids",
     "_pardon_failures": "_pardon_failures",
     "_task": "_task",
     "_plan": "_plan",
